requirements.txt
pydantic>=2.0
openai>=1.0
anthropic>=0.25
numpy>=1.24
networkx>=3.0
jinja2>=3.1
matplotlib>=3.7
python-dotenv>=1.0
